Module_1_Python_Fundamentals/exercise-01-environment-setup/sentiment-classifier/requirements.txt
# Core ML frameworks
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0

# Data processing
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

# API and serving (for future modules)
fastapi==0.104.1
uvicorn==0.24.0

# Logging and monitoring
loguru==0.7.2